use 32-bit floats

+++ src/units.rs
@@ -1,6 +1,7 @@
 /// Given a temperature in Celcius and a relative humidity percentage, returns
 /// an absolute humidity in grams/m^3.
-pub fn absolute_humidity(temp_c: f64, rel_humidity_percent: f64) -> f64 {
+// TODO(eliza): can we avoid some of the float math?
+pub fn absolute_humidity(temp_c: f32, rel_humidity_percent: f32) -> f32 {
     // first, determine the saturation vapor pressure (`P_sat`) at `temp_c`
     // degrees --- the pressure when the relative humidity is 100%. we compute
     // this using a variant of the Magnus-Tetens formula:
@@ -17,9 +18,10 @@ pub fn absolute_humidity(temp_c: f64, rel_humidity_percent: f64) -> f64 {
     // see https://carnotcycle.wordpress.com/2012/08/04/how-to-convert-relative-humidity-to-absolute-humidity/
 }
 
+
 #[cfg(test)]
 mod tests {
-    const TEST_EPSILON: f64 = 0.5; // it's an approximation lol
+    const TEST_EPSILON: f32 = 0.5; // it's an approximation lol
     use super::*;
     macro_rules! assert_float_eq {
         ($a:expr, $b:expr) => {
